migrated all tabs & search to assistantbrowse
